done checkout and order

+++ product-service/src/main/java/com/sondev/productservice/dto/request/ProductRequest.java
@@ -16,7 +16,7 @@ public class ProductRequest{
     private String description;
     private Double price;
     private Integer quantity;
-    private Integer discount;
+    private Double discount;
     private String categoryId;
     private String brandId;
     List<MultipartFile> files;
